added migration + seed method and changed int to long on personalnumber in mrgreencustomer class

+++ MRG-HomeAssignment/MRG-HomeAssignment.Lib/DomainClass/mrgreenCustomer.cs
@@ -20,6 +20,6 @@ namespace MRG_HomeAssignment.Lib.DomainClass
 
         public string City { get; set; }
 
-        public int PersonalNumber { get; set; }
+        public long PersonalNumber { get; set; }
     }
 }
